[Exchange] 코드 분할
