+++ src/components/exchange/cell/index.ts
@@ -1,2 +1,6 @@
 export { default as CurrencyInput } from './CurrencyInput';
 export { default as CurrencyTimeZone } from './CurrencyTimeZone';
+export { default as FilterTimeCell } from './FilterTimeCell';
+export { default as FilterSortCell } from './FilterSortCell';
+export { default as FilterDateCell } from './FilterDateCell';
+export { default as FilterLimitCell } from './FilterLimitCell';
